embedding_service/requirements.txt
anyio==3.7.1
astroid==2.12.12
asttokens==2.4.1
async-generator==1.10
attrs==21.4.0
Automat==22.10.0
autopep8==1.6.0
beautifulsoup4==4.9.3
blinker==1.7.0
boto3==1.18.24
botocore==1.21.24
bs4==0.0.1
cachetools==5.0.0
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.6
colorama==0.4.6
comm==0.2.1
config==0.5.1
constantly==15.1.0
cryptography==37.0.2
cssselect==1.2.0
dateparser==1.1.1
debugpy==1.8.0
decorator==5.1.1
deprecation==2.1.0
dill==0.3.2
dnslib==0.9.23
dnspython==2.4.0
docker==6.1.3
ecdsa==0.18.0
exceptiongroup==1.1.2
executing==2.0.1
filelock==3.12.2
fire==0.4.0
Flask==3.0.0
fonttools==4.38.0
fsspec==2023.12.2
google-api-core==1.31.6
google-api-python-client==2.11.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.56.3
greenlet==2.0.2
h11==0.13.0
httpcore==0.17.3
httplib2==0.20.4
huggingface-hub==0.20.2
hyperlink==21.0.0
idna==2.10
img2pdf==0.4.4
importlib-resources==5.9.0
incremental==22.10.0
ipykernel==6.28.0
ipython==8.19.0
isort==5.10.1
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.0
lazy-object-proxy==1.8.0
localstack-core==2.2.0
localstack-ext==2.2.0
lxml==4.9.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.0
opencv-python==4.6.0.66
outcome==1.2.0
packaging==21.3
parsel==1.8.1
parso==0.8.3
pbr==5.11.1
pdf2docx==0.5.6
pikepdf==6.2.1
Pillow==9.2.0
platformdirs==2.5.3
playwright==1.35.0
playwright-stealth==1.0.5
plux==1.3.2
prompt-toolkit==3.0.43
Protego==0.3.0
protobuf==3.20.1
psutil==5.9.5
pure-eval==0.2.2
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.0
pycparser==2.21
PyDispatcher==2.0.7
PyDrive==1.3.1
pyee==9.0.4
Pygments==2.15.1
pylint==2.15.5
PyMuPDF==1.21.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.20.0
python-jose==3.3.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==6.0
pyzmq==25.1.2
queuelib==1.6.2
regex==2022.3.2
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.4.2
rsa==4.8
s3transfer==0.5.2
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
scrapinghub==2.4.0
Scrapy==2.10.0
scrapy-playwright==0.0.30
selenium==4.7.0
semver==3.0.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.3.1
service-identity==23.1.0
shub==2.14.5
singleton-decorator==1.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.6.3
stevedore==5.1.0
sympy==1.12
tabulate==0.9.0
tailer==0.4.1
termcolor==2.1.1
threadpoolctl==3.2.0
tldextract==3.4.4
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.55.1
traitlets==5.14.1
transformers==4.36.2
trio==0.21.0
trio-websocket==0.9.2
Twisted==22.10.0
twisted-iocpsupport==1.0.4
typing_extensions==4.7.1
tzdata==2022.1
tzlocal==4.2
uritemplate==3.0.1
urllib3==1.26.9
w3lib==2.1.2
wcwidth==0.2.13
webdriver-manager==3.8.5
websocket-client==1.6.1
Werkzeug==3.0.1
windows-curses==2.3.1
wrapt==1.14.1
wsproto==1.1.0
yachalk==0.1.5
zope.interface==6.0
